+++ logger.d.ts
@@ -1,3 +1,3 @@
 export declare function setLogLevel(level: number): void;
 export declare function logExpression(msg: any, level?: number): void;
-export declare function setEnabled(enabled: boolean): void;
+export declare function setLoggerEnabled(enabled: boolean): void;