+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.2.0
 
-* Add new `setEnabled` function to enable / disable logger. Logger is enabled by default.
+* Add new `setLoggerEnabled` function to enable / disable logger. Logger is enabled by default.
 * Fix type file not exporting function declarations
 * Fix typo in `setLogLevel` function declaration in type file